Extract the shared logic to trait
