+++ src/Common/Crypto/CryptoGen.php
@@ -38,6 +38,8 @@ use Psr\Log\LoggerInterface;
 
 class CryptoGen implements CryptoInterface
 {
+    use CheckLatestEncryptionValueTrait;
+
     /**
      * Key cache to optimize key collection, which avoids numerous repeat
      * calls to collect the key sets (and repeat decryption of the key set
@@ -771,44 +773,5 @@ class CryptoGen implements CryptoInterface
     protected function fileExists(string $filename): bool
     {
         return file_exists($filename);
-    }
-
-    public function isDatabaseValueLatest(string $value): bool
-    {
-        if ($value === '') {
-            // Empty never encrypted
-            return true;
-        }
-
-        try {
-            $version = KeyVersion::fromPrefix($value);
-            // Value IS encrypted
-            return $this->shouldEncryptForDatabase && $version === KeyVersion::CURRENT;
-        } catch (\ValueError) {
-            // Value is NOT encrypted.
-            return !$this->shouldEncryptForDatabase;
-        }
-        // db not togglable yet
-        try {
-            return KeyVersion::fromPrefix($value ?? '') === KeyVersion::CURRENT;
-        } catch (\ValueError) {
-            return false;
-        }
-    }
-
-    public function isFilesystemValueLatest(string $value): bool
-    {
-        if ($value === '') {
-            return true;
-        }
-
-        try {
-            $version = KeyVersion::fromPrefix($value);
-            // Value IS encrypted
-            return $this->shouldEncryptForFilesystem && $version === KeyVersion::CURRENT;
-        } catch (\ValueError) {
-            // Value is NOT encrypted.
-            return !$this->shouldEncryptForFilesystem;
-        }
     }
 }
